FixingMethods.py updates pre-existing legacyidsfixer.py and introduces a new method to delete concept labels from the terms index
