+++ arches/management/commands/packages.py
@@ -35,7 +35,7 @@ from arches.app.search.search_engine_factory import SearchEngineFactory
 from arches.app.models import models
 import csv
 import arches.app.utils.backlogids as create_backlog
-import arches.app.utils.legacyidsfixer as legacy_fixer
+from arches.app.utils.FixingMethods import LegacyIdsFixer,IndexConceptFixer
 from arches.app.utils.load_relations import LoadRelations,UnloadRelations
 
 class Command(BaseCommand):
@@ -46,7 +46,7 @@ class Command(BaseCommand):
     
     option_list = BaseCommand.option_list + (
         make_option('-o', '--operation', action='store', dest='operation', default='setup',
-            type='choice', choices=['setup', 'install', 'setup_db', 'start_elasticsearch', 'setup_elasticsearch', 'build_permissions', 'livereload', 'load_resources', 'remove_resources', 'load_concept_scheme', 'index_database','export_resource_graphs','export_resources','create_backlog', 'remove_resources_from_csv', 'legacy_fixer', 'load_relations', 'unload_relations'],
+            type='choice', choices=['setup', 'install', 'setup_db', 'start_elasticsearch', 'setup_elasticsearch', 'build_permissions', 'livereload', 'load_resources', 'remove_resources', 'load_concept_scheme', 'index_database','export_resource_graphs','export_resources','create_backlog', 'remove_resources_from_csv', 'legacy_fixer', 'load_relations', 'unload_relations', 'delete_indices'],
             help='Operation Type; ' +
             '\'setup\'=Sets up Elasticsearch and core database schema and code' + 
             '\'setup_db\'=Truncate the entire arches based db and re-installs the base schema' + 
@@ -121,6 +121,8 @@ class Command(BaseCommand):
             self.load_relations(options['source'])
         if options['operation'] == 'unload_relations':
             self.unload_relations(options['source'])
+        if options['operation'] == 'delete_indices':
+            self.delete_indices(options['source'])
             
     def setup(self, package_name):
         """
@@ -375,13 +377,16 @@ class Command(BaseCommand):
                 csvwriter.writerow({k: str(v).encode('utf8') for k, v in csv_record.items()})
                 
     def legacy_fixer(self, source):
-        legacy_fixer.LegacyIdsFixer(source)
+        LegacyIdsFixer(source)
 
     def load_relations(self, source):
         LoadRelations(source)
         
     def unload_relations(self, source):
         UnloadRelations(source)
+        
+    def delete_indices(self, source):
+        IndexConceptFixer(source)
         
     def start_livereload(self):
         from livereload import Server
